Completed stateful rhythm and octave parsers
Everything compiles now, but there seems to be no way to test anything,
since the base monad in the transformer stack is no longer IO. Also,
in rhythmP and octaveP, there is an abundance of case statements to
unwrap maybes, which is highly undesireable. There is probably a way to
fix this, but we'll keep it for now, as being able to test something is
vastly more important than having the code look pretty.