Update Readme
This new Readme is the authoritative source for how the parser should behave. Currently, the parser still supports too many articulations and does not handle chords as I would like it to. Refactoring again is something of an annoying undertaking, but in my view it is worthwhile to get this right before moving on to the next phase.